Start incident data display
