+++ frontend/compositions/incident-view/incident-view-header/incident-view-header.module.css
@@ -18,6 +18,6 @@
 }
 
 .category {
-  font-size: var(--size10);
+  font-size: var(--size12);
   color: var(--grey);
 }
