+++ frontend/compositions/incident-view/incident-view-header/incident-view-header.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react"
 import IncidentViewHeader from "."
-import sampleIncident, { Incident } from '../../../helpers/sample-incident'
+import sampleIncident, { Incident } from '../../../helpers/incident'
 
 export default {
   title: "Compositions/Incident View Components/IncidentHeader"
