+++ frontend/helpers/incident.ts
